adapt for ADOL-C 2.5.2

+++ meschach/memstat.c
@@ -80,7 +80,7 @@ static unsigned int mem_hash_idx_end = 0;
 static unsigned int mem_hash(ptr)
 void **ptr;
 {
-   unsigned long lp = (unsigned long)ptr;
+   long long lp = (long long)ptr;
 
    return (lp % MEM_HASHSIZE);
 }
